refactor: :recycle: refactor and useless files removing
